Fix spacelayout to use router view
